исправил

+++ app.js
@@ -15,12 +15,7 @@ const server = http.createServer().listen(config.get('port'), config.get('ip'), 
 
 const io = require('socket.io')(server);
 
-/* const confsql = { //переложить в nconf
-    host: "localhost",
-    database: "postgres",
-    user: 'postgres',
-    password: 'sevenseven'
-}; */
+
 
 //console.log(config.get("confsql"));
 
@@ -34,23 +29,41 @@ pool.query("CREATE TABLE users1 (Id SERIAL PRIMARY KEY, Name   CHARACTER VARYING
     }
 });
 
-pool.query("CREATE  RULE rule1 AS ON INSERT TO public.users1 DO NOTIFY watchers, 'INSERT';", (err, res) => {
+pool.query("CREATE OR REPLACE RULE  rule1 AS ON INSERT TO public.users1 DO NOTIFY watchers, 'INSERT';", (err, res) => {
     if (err) {
         console.log(err.stack)
     }
 });
 
-pool.query("CREATE  RULE rule2 AS ON UPDATE TO public.users1 DO NOTIFY watchers, 'update';", (err, res) => {
+pool.query("CREATE OR REPLACE  RULE rule2 AS ON UPDATE TO public.users1 DO NOTIFY watchers, 'update';", (err, res) => {
     if (err) {
         console.log(err.stack)
     }
 });
 
-pool.query("CREATE  RULE rule3 AS ON DELETE TO public.users1 DO NOTIFY watchers, 'DELETE';", (err, res) => {
+pool.query("CREATE OR REPLACE  RULE rule3 AS ON DELETE TO public.users1 DO NOTIFY watchers, 'DELETE';", (err, res) => {
     if (err) {
         console.log(err.stack)
     }
 });
+///???
+pool.query("SELECT * FROM users1 WHERE name = 'admin'", (err, res) => {
+    if (err) {
+        console.log(err.stack)
+    }
+    if (res.rows.length == 0) {
+        //console.log("admin " + JSON.stringify(res.rows[0]));
+        pool.query("INSERT INTO users1(name, passw) VALUES ('almin', 'almin')", (err, res) => {
+
+            if (err) {
+                console.log(err.stack)
+            }
+        });
+
+    };
+});
+
+///???
 
 io.sockets.on('connection', function(client) {
     // if (client.name) {
@@ -101,8 +114,9 @@ io.sockets.on('connection', function(client) {
     client.on("PaddB", function(data) { //добавить
 
         //console.log(data);
-        if (!client.Token) {
-
+        if (client.Token) {
+            client.emit("message", "Вы уже!")
+        } else {
             pool.query('INSERT INTO users1(name, passw) VALUES($1, $2) ', data, (err, res) => {
                 // client.emit("message", "data " + data);
                 if (err) {
@@ -123,7 +137,7 @@ io.sockets.on('connection', function(client) {
             //console.log("err.stack")
             client.broadcast.emit('whuN');
 
-        } else { client.emit("message", "Вы уже!"); return };
+        };
     });
 
     /* client.on("PdelAll", function() {
@@ -139,19 +153,21 @@ io.sockets.on('connection', function(client) {
     }); */
 
     client.on("PdelB", function(data) { //удаление по имени
-        pool.query("DELETE FROM users1 WHERE  name='" + data + "'", (err, res) => {
-            if (err) {
-                console.log(err.stack)
-            }
+        if (client.name === "admin" && data !== "admin") {
+            pool.query("DELETE FROM users1 WHERE  name='" + data + "'", (err, res) => {
+                if (err) {
+                    console.log(err.stack)
+                }
 
-        })
-        console.log("err.stack")
+            })
+            console.log("err.stack")
 
-        //setTimeout(function() {
-        valueFind();
-        // }, 200);
+            //setTimeout(function() {
+            valueFind();
+            // }, 200);
 
-        client.broadcast.emit('whuN');
+            client.broadcast.emit('whuN');
+        } else client.emit("message", "Вы не АДМИН!");
     });
 
 
@@ -180,8 +196,8 @@ io.sockets.on('connection', function(client) {
                 console.log(err.stack)
             }
             console.log("нашел " + JSON.stringify(res.rows));
-            client.emit("Pserchres", res.rows);
-        })
+            client.emit("Pserchres", res.rows.length); //лучше не отправлять vse
+        });
 
 
         /*  setTimeout(function() {
@@ -193,6 +209,7 @@ io.sockets.on('connection', function(client) {
 
     client.on("login", function(data) {
         delete client.Token;
+        delete client.name;
         pool.query("SELECT * FROM users1 WHERE name = '" + data[0] + "'", (err, res) => {
                 if (err) {
                     console.log(err.stack)
@@ -204,6 +221,7 @@ io.sockets.on('connection', function(client) {
                 if (res.rows.length > 0 && data[1] && data[1] === res.rows[0].passw) {
                     client.Token = generateToken(res.rows[0].name);
                     client.name = res.rows[0].name; //!!!!!!!!
+                    client.emit("sendname", client.name);
                     client.emit("message", "удачно " + res.rows[0].passw + "=" + data[1] + " " + res.rows[0].name + "=" + data[0] + " tok=" + client.Token);
                 } else { client.emit("message", "не удачнно " + JSON.stringify(res.rows[0]) + " tok=" + client.Token); }
             }) //отправить массивом
